[CW calc] Change "lock me!" to "calculate me!"
Minor tweak, but considering the field you lock is the one where the number changes depending on the other two fields, I'd say the other two were actually the locked ones? I.e. you lock the total mark to 40% and your coursework to 51% and then it calculates the exam mark.